2/src_final/requirements.txt
astroid==1.4.9
asttokens==1.1.13
automationhat==0.2.0
blinker==1.3
blinkt==0.1.2
buttonshim==0.0.2
CacheControl==0.12.6
cachetools==4.2.4
Cap1xxx==0.1.3
certifi==2022.5.18
chardet==4.0.0
click==6.6
colorama==0.3.7
colorzero==1.1
cryptography==1.7.1
dnspython==1.16.0
docutils==0.13.1
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
firebase==4.0.1
firebase-admin==4.5.3
Flask==0.12.1
fourletterphat==0.1.0
google-api-core==1.23.0
google-api-python-client==1.12.11
google-auth==1.23.0
google-auth-httplib2==0.1.0
google-cloud-core==1.4.4
google-cloud-firestore==1.9.2
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gpiozero==1.5.0
gps==3.19
gpsd-py3==0.3.0
grpcio==1.41.1
httplib2==0.22.0
idna==2.10
isort==4.2.5
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.2
msgpack==1.0.5
mypy==0.470
numpy==1.12.1
oauthlib==2.0.1
pantilthat==0.0.7
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.5
pigpio==1.38
Pillow==4.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pygame==1.9.3
Pygments==2.2.0
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
pyliblo==0.10.0
pylint==1.6.5
pymongo==3.13.0
pynmea2==1.19.0
pyOpenSSL==16.2.0
pyparsing==2.4.7
pyperclip==1.5.27
pyserial==3.5
python-apt==1.4.3
python-firebase==1.2
pytz==2023.3
pyxdg==0.25
rainbowhat==0.1.0
requests==2.25.1
requests-oauthlib==0.7.0
roman==2.0.0
RPi.GPIO==0.7.1
rsa==4.7.2
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.1
sense-hat==2.2.0
simplejson==3.10.0
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
spidev==3.3
thonny==3.1.0
thonny-pi==1.1
touchphat==0.0.1
twython==3.4.0
typed-ast==0.6.3
unicornhathd==0.0.4
uritemplate==3.0.1
urllib3==1.26.9
Werkzeug==0.11.15
wrapt==1.9.0
